Temporary fixes for the prototype
In client, :useSSL should be an injection.

config.js should be injected from a build script, not have
configuration written openly in the git repo.

Weird advanced compilation bug in the MQTT client implementation of
the web app. Moving to goog.object fixed the problem, but reading the
code as it was, it should have worked.