Shadowed settings for not messing up with MQTT messages
When updating Teleporter A (which you control), MQTT sends messages on
the broadcast topic, so that any app listening in can update the
controls on that app instance. This causes problem for the person
using any of the sliders, as the slider will jump between various
points based on the replies from MQTT. By shadowing the the value in a
separate part of the db, we remove this problem.