Add pairing, refactor MQTT initialization
Pairing had teleporter view (status), pairing view, confirm pairing view
and unpairing view added in terms of views.

The topbar was updated to show Linked/Not linked status. This still
needs to factor in online/offline status. Are you still linked when
offline?

The MQTT initialization was refactored to use the UUIDv5 channel
representation of the auth.user/id. This is so that the platform can
easily know which channel a particular user app is listening in on.