Fix edge case on pairing and a cleanup bug
When a Teleporter is rebooted, we want a clean slate. The Teleporter
initially reports to the Platform that it is now starting up. When the
Platform receives the information that the Teleporter is now online,
it checks in the database who was paired with the Teleporter, and then
sends a message on the user MQTT topic. If the app is online, it will
catch the reboot and act accordingly. If the app is not online, the
app init will pick up on the lost pairing.