Add knob component
Component has a wheel, an arc and an overload indicator.

The wheel is turned on the x or y axis. The turning is smooth (can be
controlled with sensitivity) and switches the value depending on
rotation, rather than clicking between values.

The component can take an outside model, which then can be updated
from the outside, and the wheel and arc update correspondingly.

The overload indicator takes a model from the outside.