Add Input 1/2
Ca 70% done. All the major elements are present. It's hooked up
against re-frame (but not tested out with values from
Platform). Messages are sent over MQTT to the Teleporter with changes.

Tweaks will be needed for some of the controls, and some has such a
wide range, that both the control itself cannot handle the range, as
well as the messages sent are out of hand (think 100+ messages for
moving the control a tiny amount).