Update room functionality and style room
Share link is still not tested, or worked upon on the other end of
catching it and doing stuff with it.

Host room now appears to work.

Join room now appears to work.

All rooms styled.

When clicking on room, when you are in a room jam, leads you back to
the room, instead of going back to the jam.