Add room jam history to join room view
Room history is now shown. 5 latest only.

Some styling changes on the listing. Also added to host room list.

The history is fetched every time a join room is clicked. This should
be cached a bit more on the backend, as the frontend is reasonably
stupid (a good thing).

Footer also got an update, where it will go to the room jam if you're
'jamming'. 'jamming' here means that you either are hosting (waiting
for someone to join), knocking (waiting to join) or jamming.