Rework presets component
Based on feedback from Christian the preset now sits next to the PAN
knob. This presents some problems with how the presets are to be
presented as they are supposed to be a drop down menu. Given time
constraints the presets menu have been made into an absolute position,
that in addition is calculated based on other elements in the layout.

As such, the complexity has increased manyfold.