Add room share functionality
Added the receiving end of the room share.

The share link got changed from using a fragment (pretty URL) to using
the query params. The reason for this are the way browsers work where
the fragment is read in at a later stage in the rendering phase than
we need. Query params has no such limitations, and as such can
always (I hope) read in to determine if it's a share link or not.

Once the invite page that the share link goes to is loaded, the page
reloads without the the query parameters, in order to go back to a
pretty URL.

The logo + slogan is re-used for the initial invite page.