Fix share link catch
When a client catches a shared link, there were some edge cases when
you're not logged in. These should now be fixed.

In addition, if you're not logged in, the share link is removed from
the URL in the navigation bar when navigating to either sign up or to
login.

It's not removed in start, since it can take some time before the data
is sent back from the Platform, telling us whether we're logged in or
not, at which point a second piece of logic (the original) can take
you to the room-join page for the invite.