Force a login once you have verified your email
Due to security concerns, we force a login after having confirmed an
email address via the verify-email view. While we could technically do
a login automatically from having confirmed the email, it poses a
small window of security risk where a MITM attack could grab the
token, verify it, and gain access to the account. Or a social attack
vector ¯\_(ツ)_/¯.